@@ -29,64 +45,43 @@
     "lint-staged": "^10.5.3",
     "music-metadata-browser": "^2.5.9",
     "netmd-exploits": "^0.5.4",
-    "netmd-js": "^4.1.4",
+    "netmd-js": "^4.1.7",
     "netmd-tocmanip": "^0.1.5",
     "prettier": "^1.19.1",
-    "react": "^16.14.0",
-    "react-beautiful-dnd": "^13.1.0",
-    "react-dom": "^16.14.0",
-    "react-dropzone": "^10.2.2",
-    "react-redux": "^7.2.2",
-    "react-scripts": "3.3.1",
+    "react": "^18.2.0",
+    "react-beautiful-dnd": "^13.1.1",
+    "react-dom": "^18.2.0",
+    "react-dropzone": "^14.2.3",
+    "react-redux": "^9.1.0",
     "react95": "^3.5.0",
     "recorderjs": "^1.0.1",
     "redux-batched-actions": "^0.4.1",
-    "shazam-api": "^0.1.2",
+    "shazam-api": "^0.2.0",
     "styled-components": "^5.2.1",
-    "typescript": "^4.1.3",
+    "tss-react": "^4.9.3",
+    "vite-plugin-commonjs": "^0.10.1",
+    "vite-plugin-svgr": "^4.2.0",
     "worker-loader": "^2.0.0"
   },
-  "homepage": "https://web.minidisc.wiki/",
-  "scripts": {
-    "analyze": "source-map-explorer 'build/static/js/*.js'",
-    "start": "npm run update-version && react-scripts start",
-    "build": "npm run update-version && react-scripts build && bash -c 'cat public/extra-service-worker.js build/service-worker.js > build/final-service-worker.js'",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "predeploy": "npm run build",
-    "deploy": "gh-pages -d build",
-    "update-version": "bash -c \"printf \\\"// This file has been auto-generated. Please do not modify.\\nexport const GIT_HASH = '$(git rev-parse --short --verify master)';\\nexport const GIT_DIFF = '$(git diff | wc -l)';\\nexport const BUILD_DATE = '$(date)';\\n\\\" > src/version-info.ts\""
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "prettier"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      "last 2 chrome version"
-    ],
-    "development": [
-      "last 2 chrome version"
-    ]
-  },
-  "husky": {
-    "hooks": {
-      "pre-commit": "lint-staged"
-    }
-  },
-  "lint-staged": {
-    "src/**/*.{js,jsx,ts,tsx,json,scss,md}": [
-      "prettier --check"
-    ]
-  },
   "devDependencies": {
-    "@types/dom-mediacapture-record": "^1.0.7",
-    "@types/react-beautiful-dnd": "^13.1.1",
-    "@types/wicg-mediasession": "^1.1.3",
-    "async-mutex": "^0.2.6",
-    "gh-pages": "^2.2.0",
-    "source-map-explorer": "^2.5.3"
+    "@types/jest": "^24.0.0",
+    "@types/node": "^12.19.14",
+    "@types/react": "^18.2.43",
+    "@types/react-beautiful-dnd": "^13.1.8",
+    "@types/react-dom": "^18.2.17",
+    "@types/react-redux": "^7.1.15",
+    "@types/styled-components": "^5.1.7",
+    "@types/w3c-web-usb": "^1.0.4",
+    "@typescript-eslint/eslint-plugin": "^6.14.0",
+    "@typescript-eslint/parser": "^6.14.0",
+    "@vitejs/plugin-react": "^4.2.1",
+    "eslint": "^8.55.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-react-refresh": "^0.4.5",
+    "patch-package": "^8.0.0",
+    "source-map-explorer": "^2.5.3",
+    "typescript": "^5.2.2",
+    "vite": "^5.0.8",
+    "vite-plugin-node-polyfills": "^0.19.0"
   }
 }
