@@ -1,22 +1,22 @@
 import React, { useCallback } from 'react';
-import { useShallowEqualSelector } from '../../utils';
+import { useShallowEqualSelector } from "../../frontend-utils";
 
 import { actions as factoryProgressDialogActions } from '../../redux/factory/factory-progress-dialog-feature';
 
-import Dialog from '@material-ui/core/Dialog';
-import DialogActions from '@material-ui/core/DialogActions';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogContentText from '@material-ui/core/DialogContentText';
-import DialogTitle from '@material-ui/core/DialogTitle';
-import Slide from '@material-ui/core/Slide';
-import Button from '@material-ui/core/Button';
-import LinearProgress from '@material-ui/core/LinearProgress';
-import Box from '@material-ui/core/Box';
-import { makeStyles } from '@material-ui/core/styles';
-import { TransitionProps } from '@material-ui/core/transitions';
+import Dialog from '@mui/material/Dialog';
+import DialogActions from '@mui/material/DialogActions';
+import DialogContent from '@mui/material/DialogContent';
+import DialogContentText from '@mui/material/DialogContentText';
+import DialogTitle from '@mui/material/DialogTitle';
+import Slide, { SlideProps } from '@mui/material/Slide';
+import Button from '@mui/material/Button';
+import LinearProgress from '@mui/material/LinearProgress';
+import Box from '@mui/material/Box';
+import { makeStyles } from 'tss-react/mui';
+import { TransitionProps } from '@mui/material/transitions';
 import { useDispatch } from 'react-redux';
 
-const useStyles = makeStyles(theme => ({
+const useStyles = makeStyles()(theme => ({
     progressPerc: {
         marginTop: theme.spacing(1),
     },
@@ -26,17 +26,17 @@ const useStyles = makeStyles(theme => ({
 }));
 
 const Transition = React.forwardRef(function Transition(
-    props: TransitionProps & { children?: React.ReactElement<any, any> },
+    props: SlideProps,
     ref: React.Ref<unknown>
 ) {
     return <Slide direction="up" ref={ref} {...props} />;
 });
 
 export const FactoryModeProgressDialog = (props: {}) => {
-    const classes = useStyles();
+    const { classes } = useStyles();
     const dispatch = useDispatch();
 
-    let { visible, actionName, units, currentProgress, totalProgress, additionalInfo, canBeCancelled, cancelled } = useShallowEqualSelector(
+    const { visible, actionName, units, currentProgress, totalProgress, additionalInfo, canBeCancelled, cancelled } = useShallowEqualSelector(
         state => state.factoryProgressDialog
     );
 
@@ -44,7 +44,7 @@ export const FactoryModeProgressDialog = (props: {}) => {
         dispatch(factoryProgressDialogActions.setCancelled(true));
     }, [dispatch]);
 
-    let progressValue = Math.round((100 / (totalProgress || 1)) * currentProgress);
+    const progressValue = Math.round((100 / (totalProgress || 1)) * currentProgress);
 
     return (
         <Dialog
