@@ -1,26 +1,26 @@
 import React, { useCallback } from 'react';
 import { useDispatch } from 'react-redux';
-import { useShallowEqualSelector } from '../utils';
+import { useShallowEqualSelector } from "../frontend-utils";
 import { actions as panicDialogActions } from '../redux/panic-dialog-feature';
 
-import Dialog from '@material-ui/core/Dialog';
-import DialogActions from '@material-ui/core/DialogActions';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogTitle from '@material-ui/core/DialogTitle';
-import Slide from '@material-ui/core/Slide';
-import Button from '@material-ui/core/Button';
-import Typography from '@material-ui/core/Typography';
-import { TransitionProps } from '@material-ui/core/transitions';
-import { makeStyles } from '@material-ui/core';
+import Dialog from '@mui/material/Dialog';
+import DialogActions from '@mui/material/DialogActions';
+import DialogContent from '@mui/material/DialogContent';
+import DialogTitle from '@mui/material/DialogTitle';
+import Slide, { SlideProps } from '@mui/material/Slide';
+import Button from '@mui/material/Button';
+import Typography from '@mui/material/Typography';
+import { TransitionProps } from '@mui/material/transitions';
+import { makeStyles } from 'tss-react/mui';
 
 const Transition = React.forwardRef(function Transition(
-    props: TransitionProps & { children?: React.ReactElement<any, any> },
+    props: SlideProps,
     ref: React.Ref<unknown>
 ) {
     return <Slide direction="up" ref={ref} {...props} />;
 });
 
-const useStyles = makeStyles(theme => ({
+const useStyles = makeStyles()(theme => ({
     codeBlock: {
         marginTop: theme.spacing(3),
         fontFamily: 'monospace',
@@ -35,9 +35,9 @@ const useStyles = makeStyles(theme => ({
 
 export const PanicDialog = (props: {}) => {
     const dispatch = useDispatch();
-    const classes = useStyles();
+    const { classes } = useStyles();
 
-    let { visible, dismissed, errorProvided } = useShallowEqualSelector(state => state.panicDialog);
+    const { visible, dismissed, errorProvided } = useShallowEqualSelector(state => state.panicDialog);
 
     const handleReloadApp = useCallback(() => {
         window.location.reload();
