@@ -1,19 +1,19 @@
 import React from 'react';
-import { useShallowEqualSelector } from '../utils';
+import { useShallowEqualSelector } from "../frontend-utils";
 
-import Dialog from '@material-ui/core/Dialog';
-import DialogActions from '@material-ui/core/DialogActions';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogContentText from '@material-ui/core/DialogContentText';
-import DialogTitle from '@material-ui/core/DialogTitle';
-import Slide from '@material-ui/core/Slide';
-import LinearProgress from '@material-ui/core/LinearProgress';
-import Box from '@material-ui/core/Box';
-import { makeStyles } from '@material-ui/core/styles';
-import { TransitionProps } from '@material-ui/core/transitions';
+import Dialog from '@mui/material/Dialog';
+import DialogActions from '@mui/material/DialogActions';
+import DialogContent from '@mui/material/DialogContent';
+import DialogContentText from '@mui/material/DialogContentText';
+import DialogTitle from '@mui/material/DialogTitle';
+import Slide, { SlideProps } from '@mui/material/Slide';
+import LinearProgress from '@mui/material/LinearProgress';
+import Box from '@mui/material/Box';
+import { makeStyles } from 'tss-react/mui';
+import { TransitionProps } from '@mui/material/transitions';
 import { W95RecordDialog } from './win95/record-dialog';
 
-const useStyles = makeStyles(theme => ({
+const useStyles = makeStyles()(theme => ({
     progressPerc: {
         marginTop: theme.spacing(1),
     },
@@ -23,18 +23,18 @@ const useStyles = makeStyles(theme => ({
 }));
 
 const Transition = React.forwardRef(function Transition(
-    props: TransitionProps & { children?: React.ReactElement<any, any> },
+    props: SlideProps,
     ref: React.Ref<unknown>
 ) {
     return <Slide direction="up" ref={ref} {...props} />;
 });
 
 export const RecordDialog = (props: {}) => {
-    const classes = useStyles();
+    const { classes } = useStyles();
 
-    let { visible, trackTotal, trackDone, trackCurrent, titleCurrent } = useShallowEqualSelector(state => state.recordDialog);
+    const { visible, trackTotal, trackDone, trackCurrent, titleCurrent } = useShallowEqualSelector(state => state.recordDialog);
 
-    let progressValue = Math.round(trackCurrent);
+    const progressValue = Math.round(trackCurrent);
 
     const vintageMode = useShallowEqualSelector(state => state.appState.vintageMode);
     if (vintageMode) {
