@@ -18,7 +18,7 @@ export class RemoteAtracExportService extends DefaultFfmpegAudioExportService {
     }
 
     async encodeATRAC3({ format, loudnessTarget, enableReplayGain }: ExportParams): Promise<ArrayBuffer> {
-        let { data } = await this.ffmpegProcess.read(this.inFileName);
+        const { data } = await this.ffmpegProcess.read(this.inFileName);
 
         const payload = new FormData();
         payload.append('file', new Blob([data.buffer]), this.originalFileName);
@@ -51,14 +51,14 @@ export class RemoteAtracExportService extends DefaultFfmpegAudioExportService {
         encodingURL.searchParams.set('type', encoderFormat);
         if (loudnessTarget !== undefined) encodingURL.searchParams.set('loudnessTarget', loudnessTarget.toString());
         if (enableReplayGain !== undefined) encodingURL.searchParams.set('applyReplaygain', enableReplayGain.toString());
-        let response = await fetch(encodingURL.href, {
+        const response = await fetch(encodingURL.href, {
             method: 'POST',
             body: payload,
         });
         const source = await response.arrayBuffer();
         const content = new Uint8Array(source);
         const file = new File([content], 'test.at3');
-        let headerLength = (await getATRACWAVEncoding(file))!.headerLength;
+        const headerLength = (await getATRACWAVEncoding(file))!.headerLength;
         return source.slice(headerLength);
     }
     async encodeATRAC3Plus(parameters: ExportParams): Promise<ArrayBuffer> {
